pc-tools/requirements.txt
pyserial
tkinter>=3.5
